requirements.txt
snowflake-connector-python
pandas
pyarrow
schemachange